+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - A generic "dzrp" remote has been implemented which can connect via serial or socket connection.
 - Memory Views (debug commands "-mv", "-mvw" and "-mvd") allow additional parameter "bank=" to specify the bank to display.
 - Debug commands "-md", "-mdelta", "-ml", "-ms", "-msetb" and "-msetw" allow additional parameter to specify a bank to read from or write to.
+- New launch.json property "loadSysVars" to load zx16k or zx48 system variables at startup.
 
 # 3.7.4
 - PR #193: Fixed bug where pressing 'stop' would sometimes take 5 seconds to disconnect. (mame)